Add back dtype parameter for archives (#661)
## Description

<!-- Provide a brief description of the PR's purpose here. -->

In #639 and #643, we removed the `dtype` parameter entirely and replaced
it with `solution_dtype`, `objective_dtype`, and `measures_dtype`. I
think this simplifies things quite a bit, but it can be quite verbose,
and it also breaks backwards compatibility. In the interest of
maintaining backwards compatibility and providing a shortcut, this PR
thus adds back the `dtype` parameter. However, now it is no longer
allowed to be a dict, since we have those three parameters available.
The previous changes to the tutorials will be left as is so as to
promote usage of `solution_dtype`, `objective_dtype`, and
`measures_dtype`. I believe this is a fair balance, and most old code
should remain unaffected, as I anticipate passing a dict as the dtype is
fairly rare.

## TODO

<!-- Notable points that this PR has either accomplished or will
accomplish. -->

- [x] Update HISTORY to reflect deprecation
- [x] Add `parse_all_dtypes` util
- [x] Check coverage of dtype parsing utils
- [x] Add tests with `dtype`
- [x] Migrate archives

## Status

- [x] I have read the guidelines in

[CONTRIBUTING.md](https://github.com/icaros-usc/pyribs/blob/master/CONTRIBUTING.md)
- [x] I have linted and formatted my code with `ruff` and `ty`
- [x] I have tested my code by running `pytest`
- [x] I have added a description of my change to the changelog in
`HISTORY.md`
- [x] This PR is ready to go

+++ tests/archives/grid_archive_test.py
@@ -672,9 +672,7 @@ def test_values_go_to_correct_bin(dtype):
         dims=[10],
         ranges=[(0, 0.1)],
         epsilon=1e-6,
-        solution_dtype=dtype,
-        objective_dtype=dtype,
-        measures_dtype=dtype,
+        dtype=dtype,
     )
 
     # Values below the lower bound land in the first bin.
